requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
Django==4.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
exceptiongroup==1.1.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
